perubahan detail2

+++ database/migrations/2026_04_20_073054_create_produk_table.php
@@ -20,8 +20,8 @@ return new class extends Migration
             
             $table->string('foto')->nullable();
             $table->string('nama');
-            $table->decimal('harga_beli', 12, 2);
-            $table->decimal('harga_jual', 12, 2); 
+            $table->bigInteger('harga_beli');
+            $table->bigInteger('harga_jual');
             $table->integer('stok');
             $table->index('nama');
             $table->timestamps();
